Add BinaryStringTest and rename misnamed StringTest
- Renames tests/N98/Util/StringTest.php to tests/N98/Util/BinaryStringTest.php to match the class being tested.
- Updates class name to BinaryStringTest.
- Adds @covers annotation.
- Adds comprehensive test cases for trimExplodeEmpty, startsWith, and endsWith, including edge cases.
- Documents and tests that trimExplodeEmpty preserves array keys (including gaps).
